+++ code/TwitterCaptureMZ_GOV_PL.py
@@ -60,7 +60,7 @@ twitter_user = 'MZ_GOV_PL'
 pages_number=2
 
 # Temporarily: Data range to display when running the script
-data_range=slice(49,60,None)
+data_range=slice(55,65,None)
 # Temporarily: Max column width to display when running the script
 max_column_width=20